Create necessary resources to draw canvas to page
Added a container div to index.html which will act as the canvas
for the etch-a-sketch. Also added styles to styles.css to facilitate
drawing and positioning the container and it's child divs.

Finally, added JS to script.js which will draw the canvas "pixels"
in the container div and wrapped that in a function which is called
onload of the body index.html.

This all sets up the foundations for the etch-a-sketch project.