Add capability to "turn on" pixels
Added a class to styles.css that can be added to pixels to "turn on"
that pixel. Added some helper functions to enable the acquisition of
all pixels and adding a mouseenter event listener to them. The event
triggers a function which adds the class to turn on the pixel.