Add ability to adjust canvas resolution
To enable the functionality to adjust resolution, I've added a
button to the UI which prompts for a number between 16 and 100. The
code checks the value is valid and alerts if it isn't. If it is,
the canvas is cleared (all "pixels" are removed from DOM) and the
canvas is redrawn with new "pixels". Several helper functions have
been added to assist (calculating the "pixel" dimensions, etc) with
all these requirements.